+++ edgy_guardian/loader.py
@@ -30,7 +30,7 @@ async def handle_content_types() -> None:
     from edgy_guardian.apps import get_apps
 
     try:
-        existing_content_types: list[BaseContentType] = await get_content_type_model().query.all()
+        existing_content_types: list[BaseContentType] = await get_content_type_model().guardian.all()
     except KeyError:
         raise GuardianImproperlyConfigured(
             "EdgyGuardian requires a `content types` app/model to be installed and it seems that it is not installed or it was accidentally removed."
@@ -57,10 +57,10 @@ async def handle_content_types() -> None:
 
     for name, models in deleted_apps.items():
         for model in models:
-            await get_content_type_model().query.filter(app_label=name, model=model).delete()
+            await get_content_type_model().guardian.filter(app_label=name, model=model).delete()
 
     for name, models in new_apps.items():
         for model in models:
-            await get_content_type_model().query.get_or_create(app_label=name, model=model)
+            await get_content_type_model().guardian.get_or_create(app_label=name, model=model)
 
     logger.info("Content types have been successfully managed.")