+++ edgy_guardian/permissions/managers.py
@@ -223,18 +223,18 @@ class PermissionManager(edgy.Manager, ManagerMixin):
         ]
 
         # Bulk inserts or creates the permissions and internally Edgy does in an atomic way
-        permissions = await self.permissions_model.query.bulk_get_or_create(
+        permissions = await self.permissions_model.guardian.bulk_get_or_create(
             permissions, unique_fields=["content_type", "codename"]
         )
 
         # Make sure we add all permissions to the filter
         for perm in permissions:
-            self.permissions_model.query.filter(
+            self.permissions_model.guardian.filter(
                 codename=perm.codename, content_type=perm.content_type
             )
 
         # Get all permissions that were created
-        permissions = await self.permissions_model.query.all()
+        permissions = await self.permissions_model.guardian.all()
 
         # Assign permissions in bulk
         kwargs = {
@@ -248,7 +248,7 @@ class PermissionManager(edgy.Manager, ManagerMixin):
         """
         Checks if user has any permissions for given object.
         """
-        return await self.permissions_model.query.has_permission(user=user, perm=perm, obj=obj)
+        return await self.permissions_model.guardian.has_permission(user=user, perm=perm, obj=obj)
 
 
 class GroupManager(edgy.Manager, ManagerMixin):
@@ -306,7 +306,7 @@ class GroupManager(edgy.Manager, ManagerMixin):
         # Handles the content type for permissions assignment
         ctype = await get_content_type(obj)
         if not isinstance(perm, self.permissions_model):
-            permission, _ = await self.permissions_model.query.get_or_create(
+            permission, _ = await self.permissions_model.guardian.get_or_create(
                 content_type=ctype, codename=perm.lower(), name=perm.capitalize()
             )
         else:
@@ -373,18 +373,18 @@ class GroupManager(edgy.Manager, ManagerMixin):
         ]
 
         # Bulk inserts or creates the permissions and internally Edgy does in an atomic way
-        permissions = await self.permissions_model.query.bulk_get_or_create(
+        permissions = await self.permissions_model.guardian.bulk_get_or_create(
             permissions, unique_fields=["content_type", "codename"]
         )
 
         # Make sure we add all permissions to the filter
         for perm in permissions:
-            self.permissions_model.query.filter(
+            self.permissions_model.guardian.filter(
                 codename=perm.codename, content_type=perm.content_type
             )
 
         # Get all permissions that were created
-        permissions = await self.permissions_model.query.all()
+        permissions = await self.permissions_model.guardian.all()
 
         group_kwargs = {
             "perms": permissions,