+++ tests/guardian/guardian/tests/contents/test_content_type.py
@@ -10,28 +10,28 @@ models = settings.edgy_guardian.registry
 
 
 async def test_get_for_model(client):
-    content_type = await ContentType.query.get_for_model("users")
+    content_type = await ContentType.guardian.get_for_model("users")
 
     assert content_type is not None
 
 
 async def test_get_for_id(client):
-    content_type = await ContentType.query.get_for_id(1)
+    content_type = await ContentType.guardian.get_for_id(1)
 
     assert content_type is not None
 
 
 async def test_cache(client):
-    content_type = await ContentType.query.get_for_id(1)
+    content_type = await ContentType.guardian.get_for_id(1)
 
     assert content_type is not None
 
-    assert ContentType.query._cache is not None
+    assert ContentType.guardian._cache is not None
 
-    ContentType.query.clear_cache()
+    ContentType.guardian.clear_cache()
 
-    assert ContentType.query._cache == {}
+    assert ContentType.guardian._cache == {}
 
-    content_type = await ContentType.query.get_for_id(1)
+    content_type = await ContentType.guardian.get_for_id(1)
 
     assert content_type is not None