+++ docs/release-notes.md
@@ -2,12 +2,17 @@
 
 ## 0.1.1
 
+### Changed
+
+- Rename internal models manager default from `query` to `guardian`. This will avoid any clash or
+the existing `query` override in any edgy models.
+
 ### Fixed
 
 - Unnecessary load of apps for ContentTypes.
 - Mandatory obj attribute in `assign_perm`.
 - Mypy typing and checks
-  
+
 ## 0.1.0
 
 ### Added